Sponsors
Big update creates a custom post type for sponsors.

Additionally, a drop-down is added to posts and sessions so that a
sponsor can be added to those pages; if so, a banner or shoulderbox are
added, respectively.

There are now four sidebars that display a row of logos at specific
places, based on a similar dropdown menu; if it is left to 'none', then
a "Your logo here" box displays.

Multiple image sizes are created for featured image on sponsor pages.